chore: guard accumulator realloc and clean up build warnings
Pre-release sanitization from the staleness/bug sweep. No behavior change.
Verified: clean debug (scalar) and release (AVX2, -j) builds — zero warnings
in our code — and lint clean.

#1 async_sample_callback: the accumulator-grow branch acquired a same-pool
   frame (capacity = FFT size) and resized it to new_total. FFTFrame::resize
   only sets the logical size over a fixed-capacity block and its bounds assert
   is compiled out under NDEBUG, so new_total > capacity would silently overrun
   the heap in release. It is currently unreachable — buf_len (32768 complex
   samples) is an exact multiple of every supported FFT size, so each device
   chunk is exactly fft_size and the accumulator always drains to empty — but a
   future buf_len or FFT-size change could break that. Added a capacity check
   that drops the buffer instead of overrunning, plus a comment recording the
   invariant. (Dropped the now-dead std::max with g_async_fft_size: inside this
   branch new_total always exceeds one frame.)

#2 Move the interleaved-float `data`/`win`/`src`/`dst` pointer declarations
   inside the `#ifdef __AVX2__` blocks in FftAnalyzer::execute (DC-center fill)
   and SignalProcessor::apply_window / remove_dc. They were only used on the
   AVX2 path, so scalar builds (debug, no -march=haswell) warned them unused.
   remove_dc keeps a single declaration shared by its two AVX2 blocks (same
   function scope) to avoid a redeclaration in the vectorized build.

#3 SdlRenderer::draw_left_axes: make the int->float glyph-height conversion in
   the label-clamp explicit (static_cast<float>(h)), matching the other uses of
   h on the surrounding lines, to clear a release -Wconversion warning.

+++ src/gui/sdl_renderer.cpp
@@ -522,7 +522,7 @@ void SdlRenderer::draw_left_axes() {
     int h = 0;
     m_text_renderer->get_text_size(l.text, &w, &h);
     float ty = l.y - static_cast<float>(h) / 2.0F; // center on the tick
-    ty = std::clamp(ty, 0.0F, static_cast<float>(m_height) - h);
+    ty = std::clamp(ty, 0.0F, static_cast<float>(m_height) - static_cast<float>(h));
     SDL_SetTextureBlendMode(t, SDL_BLENDMODE_BLEND);
     SDL_FRect const d = {static_cast<float>(kPad), ty, static_cast<float>(w),
                          static_cast<float>(h)};
